requirements.txt
Package                      Version            
---------------------------- -------------------
absl-py                      1.0.0              
anyio                        3.5.0              
AnyQt                        0.1.0              
asttokens                    2.0.5              
astunparse                   1.6.3              
backcall                     0.2.0              
baycomp                      1.0.2              
Bottleneck                   1.3.4              
CacheControl                 0.12.10            
cachetools                   5.0.0              
certifi                      2021.10.8          
cffi                         1.15.0             
chardet                      4.0.0              
charset-normalizer           2.0.12             
commonmark                   0.9.1              
cryptography                 36.0.2             
cycler                       0.11.0             
debugpy                      1.6.0              
decorator                    5.1.1              
dictdiffer                   0.9.0              
docutils                     0.18.1             
entrypoints                  0.4                
et-xmlfile                   1.1.0              
executing                    0.8.3              
flatbuffers                  2.0                
fonttools                    4.31.2             
gast                         0.5.3              
google-auth                  2.6.2              
google-auth-oauthlib         0.4.6              
google-pasta                 0.2.0              
grpcio                       1.44.0             
h11                          0.12.0             
h5py                         3.6.0              
httpcore                     0.13.7             
httpx                        0.19.0             
idna                         3.3                
importlib-metadata           4.11.3             
ipykernel                    6.9.2              
ipython                      8.1.1              
ipython-genutils             0.2.0              
jedi                         0.18.1             
jeepney                      0.7.1              
joblib                       1.1.0              
jupyter-client               7.1.2              
jupyter-core                 4.9.2              
keras                        2.8.0              
Keras-Preprocessing          1.1.2              
keyring                      23.5.0             
keyrings.alt                 4.1.0              
kiwisolver                   1.4.0              
libclang                     13.0.0             
lockfile                     0.12.2             
Markdown                     3.3.6              
matplotlib                   3.5.1              
matplotlib-inline            0.1.3              
msgpack                      1.0.3              
nest-asyncio                 1.5.4              
networkx                     2.7.1              
numpy                        1.22.3             
oauthlib                     3.2.0              
openpyxl                     3.0.9              
openTSNE                     0.6.2              
opt-einsum                   3.3.0              
orange-canvas-core           0.1.25             
orange-widget-base           4.16.1             
Orange3                      3.31.1             
packaging                    21.3               
pandas                       1.4.1              
parso                        0.8.3              
pexpect                      4.8.0              
pickleshare                  0.7.5              
Pillow                       9.0.1              
pip                          20.0.2             
pkg-resources                0.0.0              
primer3-py                   0.6.1              
prompt-toolkit               3.0.28             
protobuf                     3.19.4             
psutil                       5.9.0              
ptyprocess                   0.7.0              
pure-eval                    0.2.2              
pyasn1                       0.4.8              
pyasn1-modules               0.2.8              
pycparser                    2.21               
Pygments                     2.11.2             
pyparsing                    3.0.7              
PyQt5                        5.15.6             
PyQt5-Qt5                    5.15.2             
PyQt5-sip                    12.9.1             
pyqtgraph                    0.12.4             
PyQtWebEngine                5.15.5             
PyQtWebEngine-Qt5            5.15.2             
python-dateutil              2.8.2              
python-louvain               0.16               
pytz                         2022.1             
PyYAML                       6.0                
pyzmq                        22.3.0             
qasync                       0.23.0             
qtconsole                    5.2.2              
QtPy                         2.0.1              
requests                     2.27.1             
requests-oauthlib            1.3.1              
rfc3986                      1.5.0              
rsa                          4.8                
scikit-learn                 1.0.2              
scipy                        1.8.0              
SecretStorage                3.3.1              
serverfiles                  0.3.1              
setuptools                   44.0.0             
six                          1.16.0             
sklearn                      0.0                
sniffio                      1.2.0              
stack-data                   0.2.0              
tensorboard                  2.8.0              
tensorboard-data-server      0.6.1              
tensorboard-plugin-wit       1.8.1              
tensorflow                   2.8.0              
tensorflow-io-gcs-filesystem 0.24.0             
termcolor                    1.1.0              
tf-estimator-nightly         2.8.0.dev2021122109
threadpoolctl                3.1.0              
tornado                      6.1                
traitlets                    5.1.1              
typing-extensions            4.1.1              
urllib3                      1.26.9             
wcwidth                      0.2.5              
Werkzeug                     2.0.3              
wheel                        0.37.1             
wrapt                        1.14.0             
xlrd                         2.0.1              
XlsxWriter                   3.0.3              
zipp                         3.7.0              
